Update to bevy 0.11

+++ build/wasm_build.sh
@@ -6,10 +6,36 @@ set -e
 
 OutDir=target/wasm_package
 
+ReleaseFlags="--release"
+BinDirectory="release"
+RunWasmopt=1
+Target="$OutDir"
+
+while true ; do
+  case "$1" in
+    --debug)
+      ReleaseFlags=""
+      BinDirectory="debug"
+      RunWasmopt=0
+      shift
+    ;;
+    --target)
+      OutDir="$2"
+      shift 2
+    ;;
+    --*)
+      echo "ERROR: unknown option $1"
+      exit 1
+    ;;
+    *)
+      break
+    ;;
+  esac
+done
 
 if [ ! -e .git ]; then
-	echo "Must be run from repository root"
-	exit 1
+  echo "Must be run from repository root"
+  exit 1
 fi
 
 #
@@ -27,14 +53,14 @@ if [ ! -e target ] ; then
     mkdir target
 fi
 
-cargo build --release --target wasm32-unknown-unknown 
+RUSTFLAGS='--cfg=web_sys_unstable_apis' cargo build $ReleaseFlags --target wasm32-unknown-unknown 
 
 FileName="$ProjName.wasm"
-WasmFile="$(cargo metadata --format-version 1 | sed -n 's/.*"target_directory":"\([^"]*\)".*/\1/p')/wasm32-unknown-unknown/release/$FileName"
+WasmFile="$(cargo metadata --format-version 1 | sed -n 's/.*"target_directory":"\([^"]*\)".*/\1/p')/wasm32-unknown-unknown/$BinDirectory/$FileName"
 
 if [ ! -e "$WasmFile" ]; then
-	echo "Script is borken, it expects file to exist: $WasmFile"
-	exit 1
+  echo "Script is borken, it expects file to exist: $WasmFile"
+  exit 1
 fi
 
 [ ! -e "$OutDir" ] || rm -r "$OutDir"
@@ -47,22 +73,23 @@ if [ ! -e "$BINDGEN_EXEC_PATH" ] ; then
 fi
 
 $BINDGEN_EXEC_PATH \
-	--no-typescript \
-	--out-dir "$OutDir" \
-	--target web \
-	"$WasmFile"
+  --no-typescript \
+  --out-dir "$OutDir" \
+  --target web \
+  "$WasmFile"
 
 
-if [ -e $(which wasm-opt) ] ; then
-	BindgenOutput="$OutDir/${ProjName}_bg.wasm"
-	wasm-opt -Oz -o "$BindgenOutput.post-processed" "$BindgenOutput"
-	echo "Applying wasm-opt optimizations"
-	echo "before: $(wc -c "$BindgenOutput")"
-	echo "after : $(wc -c "$BindgenOutput.post-processed")"
-	mv "$BindgenOutput.post-processed" "$BindgenOutput"
-else
-	echo "Continuing without wasm-opt, it is highly recommended that you"
-	echo "install it, it has been known to divide by two wasm files size"
+if [[ -e $(which wasm-opt) && $RunWasmopt == 1 ]] ; then
+  BindgenOutput="$OutDir/${ProjName}_bg.wasm"
+  echo "Applying wasm-opt optimizations"
+  echo "before: $(wc -c "$BindgenOutput")"
+  wasm-opt -Oz \
+        --output "$BindgenOutput.post-processed" "$BindgenOutput"
+  echo "after : $(wc -c "$BindgenOutput.post-processed")"
+  mv "$BindgenOutput.post-processed" "$BindgenOutput"
+elif [[ ! $RunWasmopt ]] ; then
+  echo "Continuing without wasm-opt, it is highly recommended that you"
+  echo "install it, it has been known to divide by two wasm files size"
 fi
 
 cp build/wasm_build.html "$OutDir/index.html"
@@ -70,17 +97,12 @@ cp build/wasm_build.html "$OutDir/index.html"
 # Rename JS files
 Count=0
 for _ in $OutDir/*.js; do
-	((Count+=1))
+  ((Count+=1))
 done
 
 if [ $Count -ne 1 ]; then
-	echo "Script is broken, must be 1 JS file matching mask"
-	exit 1
+  echo "Script is broken, must be 1 JS file matching mask"
+  exit 1
 fi
 
 mv $OutDir/*.js "$OutDir/main.js"
-
-if [ "$1" ] ; then
-	trash-put "$1"
-	mv "$OutDir" "$1"
-fi
